refactor: rename ConfigPage directory to follow new naming conventions
