+++ client/src/App/App.tsx
@@ -2,7 +2,7 @@ import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import DashboardContainer from "../pages/FunctionPerformance/FunctionPerformance";
 import CloudwatchContainer from "../pages/CloudWatchMetrics/CloudWatchMetrics";
 import NavbarComponent from "../components/Navbar/Navbar";
-import ConfigPageContainer from "../pages/ConfigPage/Config";
+import ConfigPageContainer from "../pages/Config/Config";
 import ChatContainer from "../pages/FunctionPerformance/Chat/Chat";
 import "../utils/chartSetup";
